bot/requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
pillow>=10.0.0  # For image processing
opencv-python>=4.8.0  # For advanced image comparison
psycopg2-binary>=2.9.9  # PostgreSQL adapter
redis>=5.0.1  # For caching
pyjwt>=2.8.0  # For JWT handling
python-Levenshtein>=0.23.0  # For string similarity
requests>=2.31.0  # For HTTP requests
imagehash>=4.3.1  # For image similarity detection
aiohttp>=3.9.1  # For async HTTP requests
psutil>=5.9.0  # For system monitoring
sqlalchemy>=2.0.25